Исправления ошибок

+++ apps/api/src/app/vk/module/chat/chat.utils.vk.ts
@@ -46,5 +46,7 @@ export async function checkMuteList(chat: Chat): Promise<void> {
 }
 
 export async function deleteAntispam(chat: Chat): Promise<void> {
-  await AntispamModule.deleteMany({ chatId: chat.chatId, date: { $lt: moment().startOf('day').toDate() } });
+  if (chat) {
+    await AntispamModule.deleteMany({ chatId: chat.chatId, date: { $lt: moment().startOf('day').toDate() } });
+  }
 }
