+++ apps/api/src/app/vk/module/marriage/marriage.utils.vk.ts
@@ -7,14 +7,16 @@ import { ContextDefaultState, MessageContext } from "vk-io";
 import { MessagesConversationMember, UsersUserFull } from "vk-io/lib/api/schemas/objects";
 
 export async function checkTimeMarriage(chat: Chat, members: { id: number, item: MessagesConversationMember, profile: UsersUserFull, info: User }[], message: MessageContext<ContextDefaultState>) {
-  const marriages: Marriage[] = await MarriageModule.find({ chatId: chat.chatId, isConfirmed: false });
-  for (const marriage of marriages) {
-    if (!marriage.checkDate || moment().diff(moment(marriage.checkDate), 'minutes') > 0) {
-      await MarriageModule.deleteOne({ chatId: chat.chatId, userFirstId: marriage.userFirstId, userSecondId: marriage.userSecondId });
-      let result = `${await stringifyMention({ userId: marriage.userFirstId, userInfo: members.find((m) => m.id === marriage.userFirstId)?.profile })}`;
-      result = result.concat(` и ${await stringifyMention({ userId: marriage.userSecondId, userInfo: members.find((m) => m.id === marriage.userSecondId)?.profile })}`);
-      result = result.concat(` не смогли вовремя расписаться`);
-      await message.send(result).catch(console.error);
+  if (chat) {
+    const marriages: Marriage[] = await MarriageModule.find({ chatId: chat.chatId, isConfirmed: false });
+    for (const marriage of marriages) {
+      if (!marriage.checkDate || moment().diff(moment(marriage.checkDate), 'minutes') > 0) {
+        await MarriageModule.deleteOne({ chatId: chat.chatId, userFirstId: marriage.userFirstId, userSecondId: marriage.userSecondId });
+        let result = `${await stringifyMention({ userId: marriage.userFirstId, userInfo: members.find((m) => m.id === marriage.userFirstId)?.profile })}`;
+        result = result.concat(` и ${await stringifyMention({ userId: marriage.userSecondId, userInfo: members.find((m) => m.id === marriage.userSecondId)?.profile })}`);
+        result = result.concat(` не смогли вовремя расписаться`);
+        await message.send(result).catch(console.error);
+      }
     }
   }
 }
